Integrating two more App Delegate leaves
[skip ci]
