+++ Client/Ecosia/AppDelegate/Leafs/UserActivityRestoringAppDelegate.swift
@@ -1,0 +1,64 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import CoreSpotlight
+
+final class UserActivityRestoringAppDelegate: AppDelegateLeaf {
+    
+    private var browserViewController: BrowserViewController?
+    
+    required init(browserViewController: BrowserViewController?) {
+        self.browserViewController = browserViewController
+    }
+    
+    func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void) -> Bool {
+        
+        guard let browserViewController else {
+            return false
+        }
+        
+        if userActivity.activityType == SiriShortcuts.activityType.openURL.rawValue {
+            browserViewController.openBlankNewTab(focusLocationField: false)
+            return true
+        }
+
+        // If the `NSUserActivity` has a `webpageURL`, it is either a deep link or an old history item
+        // reached via a "Spotlight" search before we began indexing visited pages via CoreSpotlight.
+        if let url = userActivity.webpageURL {
+            let query = url.getQuery()
+
+            // Check for fxa sign-in code and launch the login screen directly
+            if query["signin"] != nil {
+                // bvc.launchFxAFromDeeplinkURL(url) // Was using Adjust. Consider hooking up again when replacement system in-place.
+                return true
+            }
+
+            // Per Adjust documentation, https://docs.adjust.com/en/universal-links/#running-campaigns-through-universal-links,
+            // it is recommended that links contain the `deep_link` query parameter. This link will also
+            // be url encoded.
+            if let deepLink = query["deep_link"]?.removingPercentEncoding, let url = URL(string: deepLink) {
+                browserViewController.switchToTabForURLOrOpen(url)
+                return true
+            }
+
+            browserViewController.switchToTabForURLOrOpen(url)
+            return true
+        }
+
+        // Otherwise, check if the `NSUserActivity` is a CoreSpotlight item and switch to its tab or
+        // open a new one.
+        if userActivity.activityType == CSSearchableItemActionType {
+            if let userInfo = userActivity.userInfo,
+                let urlString = userInfo[CSSearchableItemActivityIdentifier] as? String,
+                let url = URL(string: urlString) {
+                browserViewController.switchToTabForURLOrOpen(url)
+                return true
+            }
+        }
+
+        return false
+    }
+
+    
+}
